move run_sgcollect
